clean up how we organize code
