+++ busy_beaver/apps/github_integration/summary/workflow.py
@@ -1,3 +1,4 @@
+from datetime import timedelta
 import logging
 import random
 from typing import List
@@ -6,9 +7,23 @@ from sqlalchemy import and_
 
 from .blocks import GitHubSummaryPost
 from busy_beaver.common.wrappers import SlackClient
-from busy_beaver.models import GitHubSummaryUser
+from busy_beaver.exceptions import ValidationError
+from busy_beaver.extensions import rq
+from busy_beaver.models import GitHubSummaryUser, SlackInstallation
+from busy_beaver.toolbox import set_task_progress, utc_now_minus
 
 logger = logging.getLogger(__name__)
+
+
+@rq.job
+def post_github_summary_message(workspace: str):
+    installation = SlackInstallation.query.filter_by(workspace_id=workspace).first()
+    if not installation:
+        raise ValidationError("workspace not found")
+
+    boundary_dt = utc_now_minus(timedelta(days=1))
+    fetch_github_summary_post_to_slack(installation, boundary_dt)
+    set_task_progress(100)
 
 
 def fetch_github_summary_post_to_slack(installation, boundary_dt):
